Looking Into Rotation
Doesn't seem to be a single centralized method for it. Can't tell if one of the lines in AppClass actually doesn't matter or if we've just temporarily commented out the code it's for. My SDK version is different, so even though I changed it back you might have to rebuild.